Started building new concept framework for towns using inheritance.
This is !very much! a beginning but I hope to make more progress as
I learn more about how to use templates. Basically each town should have a custom template
based on its name. This template should be loaded and rendered to get the town variables
like the number and type of shops and which items each shop has.

Each town should have a separate "display" function to render the town data into
the html that is needed to load it into the main website.