Implemented new validate function as a method of the EasyDatabase class.
This is implemented in the login() function in the app class and may break everything ... I think it works so far though?
I am starting to have a problem global imports i.e. `from xxxx import *` resulting in fail do to to recursion.

Can someone explain session_id and user_id to me please?

I found a way to hide all of the SQL. I will try and implement it throughout the EasyDatabase class.