Started to build some unit tests for the EasyDatabase class.
I did this because the code was becoming so complicated that I didn't know if it was actually working or not.
The basic framework should be able to work for all the modules and will help a lot.
Before doing a commit or push just >python test_all.py
This will verify that the changes you made didn't break anything.