Rebuild locations.py
Currently bugtesting, run python locations.py to import the test suite.
It fails to load. I have no idea why. I don't understand the error.
It sounds like the name if my BaseList table is invalid but this is not the case.
It is some kind of ... I added this twice problem.