Restructured the Location classes.
Now World_Map does not derive from Location, instead they have the same root of
Base_Location.

    Oh! And always delete the test database before trying to run new code
or you will get obscure errors like locations.world_map_id does not exist while clearly it does.
It is because the database is out of date and needs to be rebuilt (delete).