Designed a Dictionary Database class.
Not fully implemented and needs to be tested fully in game_tests.py.

Plus (and this is really cool) I started work on a generic Database Object builder
for any python object. It is in marlens_scratchpad.py. And needs recursion and a lot more
work. But the bones are there :D