Made hero.primary_attributes actually saveable.
The init setup may need to be fixed?
I have found out that I need to upgrade my test suite for locations and database.
The problem is that using a session query from the session that commited the data will
return the same object ... rather than pulling the data from the database ... which
confuses the shit out of me because I thought it didn't do that?