Fix Ability page. Fix cmd button in Firefox? Plus partial debug of level up function.
More testing will be required to prove if this is true.
Found some kind of bug in level up that updates the wrong attributes.
Disabled CMD code for level up as it was conflicting with the new JavaScript code.
Ability page needs a switch to check different known abilities. Currently
defaults to Basic.

Moved CMD button code to the end of the home.html where it belongs,
plus found a slight variant that works in Firefox. I may need to change the code in
main.html too.