Set up basic generated attribute.py file that will replace PrimaryAttribute class.
I used Jinja2 to generate this new file. The generated file is the
one that will be run by the main game code. This file will be read only at
some point. Next I need to set up the generated code so that it
loads all the correct modules and is loaded by game.py (and where ever else is needed.)

The relationships will be complicated and are not correct yet. I will need to configure
the joins; see http://docs.sqlalchemy.org/en/latest/orm/join_conditions.html