Fix #192
User was not imported from game.py. This is because I have been trying to keep database code in the database.py file. (session, query etc.)
I try and add a new method to EZDB when I want to get something from the database.
I would recommend adding a

get_all_users(self):
   return self.session.query(User).order_by(User.id).all()

to the EZDB class in database.py