Inbox relationship fix.
Plus hopefully enough documentation to use the new object.
After much consideration I realized that an Inbox is really
a type of Association object ... sort of an intermediary
between a User and a Message. The true relationship is between
A User and a Message ... but this is "wrapped" by the Inbox object
which provides extra data an methods.