Found way to intercept url calls.
Now prevent_url_typing code is ignored during setup
and only runs when you try to access a url.

I found the variable that corresponds to the current url that is being requested! "request.path"

Next I need to re-write the location.py code so that each location has a url and a list of adjacent locations.
This will allow me to add hero.current_location.url and hero.current_location.adjacent_locations urls to the list of valid_urls.