Basic working location object.
Passes the initialization tests and some basic connection tests. I have adjacencies! They can be in or out but default to bidirectional.
I am working on setting up 'places_of_interest', which should be auto generated.

I am moving relationships from complex relationship into the files of the objects involved. This is a work in progress.