Final port of old locations to new.
This will need more work ... but it is less straight forward as many of the places need to be reimagined. I believe that a bunch of places can be combined and made more generic. I would also like to have the Hero's current location update whenever they move anywhere i.e. the tavern is a location ... when you go there it should update the Hero's current_location.

I also want to switch more to the "Game engine" model.