Fix build_code.py overwriting modified files!
Fix profile_poficiencies_template!
The build_code is now in one place and handles backups.
If you modify a code the is generated from a template the file will be backed up. If there is a backup it will not be overwritten. Every so often we should rebuild the templates to reflect the backups and then delete them.