Very basic working Event code!
Now I need to test, comment and add features.
And add in more Triggers and content.
And figure out how the quest popup HTML/JS/Python? code works.

And I found a bug where I can't go back to the Town once I am in the Blacksmith? Which is a pain for testing.