Fix data lost in build_code.cmd update.
NOTE: All changes in *.bak files need to be propagated back into the *_template and *_data files as these changes are part of the _current working code_ ... and are thus integral the functioning of the game.
Just make sure when you run build_code.cmd that you add everybody's changes to the template code. Since it is much easier to just do a minor tweak here and there and not use the template code ... that is what everyone does ... and the template code gets out of date.