Add in new 'specifications folder'.
This will be like a README for what much of the code in the game is _supposed_ to do. It will probably be chronically out of date or just plain wishful thinking.
On the other hand, I hope that it will provide me with some structure and a way to trace out and decribe 'code paths' (i.e. the path that data flows to execute some kind of sensible result)
Yay! I have a working event/trigger analysis.