Partially hacked together Handler decorator.
I can't figure out how to overload the __init__ which I though was going to be easy! I mean I know it is easy because I do it all the time with inheritance. But I can't use inheritance because of the whole table lameness (where SQLAlchemy inheritance has a tonne of extra complexity).