New QuestPath/Quest concept. Much more streamlined.
I will have to deal with the templating problems and rebuild the trigger/handler system.
I will also need to link the Journal to the Hero and the Journal to the QuestPath.
Debugging will be a very long process as I have written all this code raw (with no testing/running of the code).