Add in new Navigator object.
This should allow more complex Hero location movement.
The current situation is that the Location object can't handle the complexity of tasks that we are asking it to do. Mainly we need a location with 2 parents ... because it is kind of generic ... but can't work out how to get to and from this new type of location.