Reformat Inventory JS make the tootip clickable.
I put most of the items back in place but realized that I deleted a few @Elthran added. I need to add more types of slots to the Inventory code.
I need a custom handler for "rings" ... sort of a mini inventory? Probably will work like the Unequipped section.
I need to fix the JS/Command code so that you can "replace" one item with another.