Dynamic notification button.
To use set the @set_notification_active decorator above
a given command function.

To use it in other places ... I haven' built.
Though as long as the page reloads it should still work correctly.

But if the page doesn't reload you complete a quest but not using the command code it won't work.