Bug fix forum with new relationships.
Bug fix some formatting in forum.
Add in Mixin for human readable datetime formatting.

Remove use of "username.title()" as this may confuse users as there username won't be the same as it really is. Consider adding in a Nickname to user class?

Currently User ins't directly linked via a relationship to Forum.