Fix database naming. Fix detached instance bug.
If you want to change the database name you need to create a
database of the given name (in pythonanywhere gui) then delete it in
mysql console ... then start the app referencing the now missing
database.

Don't use @scoped_session on EZDB function that return a database object
as you will get a detached instance error.