Add in initial (broken?) code from Corentin Wallez.
From their blog https://blog.kangz.net/posts/2016/08/31/code-generation-the-easier-way/.
I indend to fix it up and rebuild it in such a way that it will work for me. Maybe I'll even make it a python module installable by 'pip'. Since It seems as though I am not the only person with this problem.