Fixes #317.
Additionally it might fix other percent bugs. The current value for hero.get_summed_proficiencies('health') was using the current value of the generic Health object ... not the hero.base_proficiencies.health.current value.

I really need to write a better version of that code ... it is the lack of DRYness that has caused most of the bugs in it. It was correctly implemented in lower down in the all proficiencies code .. just not in the single proficiency code.