Fix bug in page turn assignment to onclick function.
The main problems were 'event' needed to be a variable that was passed into the inner function at execution time.
The second problem was that 'response' wasn't defined in the local scope so it was garbage collected? before the onclick function was executed. It was defined when the function as assigned ... but not when the button was clicked on. I made a local variable and that seemed to work.

In python it would have been
```python
spell_id = 5
spell_img.onclick = lambda event, spell_id=spell_id: sendToPy(event, None, "cast_spell", {"id": spell_id})
```

The use of "use strict"; is cosmetic.
My change in 'commands.py' is cosmetic also. Since the variable is never used .. it could be a string or an integer. If it was going to be used ... it should be an integer.